Send and receive response from tracker

+++ TorrentClient/FileInfo.cs
@@ -2,17 +2,17 @@ namespace TorrentClient;
 
 public class FileInfo
 {
-    public int Length { get; private set; }
+    public long Length { get; private set; }
     public string? Md5Sum { get; private set; }
     public List<string>? Path { get; private set; }
 
-    public FileInfo(int length, string? md5sum = null)
+    public FileInfo(long length, string? md5sum = null)
     {
         Length = length;
         Md5Sum = md5sum;
     }
 
-    public FileInfo(int length, List<string> path, string? md5sum = null)
+    public FileInfo(long length, List<string> path, string? md5sum = null)
     {
         Length = length;
         Path = path;
